+++ package.json
@@ -28,7 +28,7 @@
     "jquery-ui": "1.12.1",
 	"moment": "2.18.1",
     "moment-timezone": "^0.5.13",
-    "ng2-bootstrap": "^1.6.3",
+    "ngx-bootstrap": "^1.6.5",
     "prismjs": "^1.6.0",
     "rxjs": "^5.3.0",
     "source-sans-pro": "^2.0.10",
